+++ include/OrderBook.h
@@ -11,17 +11,17 @@
 
 class OrderBook {
     private:
-        std::map<int64_t, std::list<Order>> buys;
-        std::map<int64_t, std::list<Order>> sells;
-        std::unordered_map<int64_t, std::tuple<int64_t, std::list<Order>::iterator>> order_lookup;
+        std::map<long long, std::list<Order>> buys;
+        std::map<long long, std::list<Order>> sells;
+        std::unordered_map<long long, std::tuple<long long, std::list<Order>::iterator>> order_lookup;
         TradeLog trade_log;
     public:
         OrderBook();
-        int64_t add_limit_order(bool isBuy, int64_t priceTick, int quantity, int64_t timestamp);
-        std::map<int64_t, std::list<Order>>::reverse_iterator get_best_bid();
-        std::map<int64_t, std::list<Order>>::iterator get_best_ask();
-        int add_IOC_order(bool isBuy, int quantity, int64_t timestamp);
-        int cancel_order(int64_t orderId);
-        void modify_order(int64_t order_id, int new_quantity, int64_t timestamp);
+        long long add_limit_order(bool isBuy, long long priceTick, int quantity, long long timestamp);
+        std::map<long long, std::list<Order>>::reverse_iterator get_best_bid();
+        std::map<long long, std::list<Order>>::iterator get_best_ask();
+        int add_IOC_order(bool isBuy, int quantity, long long timestamp);
+        int cancel_order(long long orderId);
+        void modify_order(long long order_id, int new_quantity, long long timestamp);
         void snapshot();
 };
