+++ include/Trade.h
@@ -3,13 +3,13 @@
 #include "Order.h"
 
 struct Trade {
-    int64_t tradeId;
-    int64_t priceTick;
+    long long tradeId;
+    long long priceTick;
     int quantity;
-    int64_t buyOrderId;
-    int64_t sellOrderId;
-    int64_t timestampUs;
+    long long buyOrderId;
+    long long sellOrderId;
+    long long timestampUs;
 
     Trade();
-    Trade(int64_t buyId, int64_t sellId, int64_t priceTick, int quantity, int64_t timestampUs);
+    Trade(long long buyId, long long sellId, long long priceTick, int quantity, long long timestampUs);
 };
