+++ include/TradeLog.h
@@ -5,6 +5,6 @@ class TradeLog {
         std::list<Trade> trades;
     public:
         TradeLog();
-        int64_t add_trade(int64_t buyId, int64_t sellId, int64_t priceTick, int quantity, int64_t timestampUs);
+        long long add_trade(long long buyId, long long sellId, long long priceTick, int quantity, long long timestampUs);
         void show_trades();
 };
